requirements.txt
asttokens==3.0.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.2
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
stack_data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.14
wheel==0.45.1
zipp==3.23.0
